Lazylinks: Grid initialisation with no register read-back
The aim of this is to avoid using XLB_ORIGIN_ID as a temporary node ID. This
prevents routing issues for local traffic on the boot core (such as the
ethernet stack). Of course, it also means we can't tell if something broke.

If this works (still needs to be tested), then in theory I could do something
smarter to the initial routing tables so that I can use the final peripheral
ID rather than XLB_ORIGIN_ID... if it's worth the effort, which it might be,
as detecting an error rather than hanging is much easier to handle - we can
just reset the boards again.

Let's see what happens...